requirements.txt
fastapi
uvicorn
requests
validators
langchain
langchain-core
langchain-community
langchain-huggingface
langchain-groq
youtube-transcript-api
yt-dlp
pypdf
beautifulsoup4
groq
python-multipart
streamlit
transformers
sentence-transformers
faiss-cpu
python-dotenv
